# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/kayit": {
            "post": {
                "tags": [
                    "Base API"
                ],
                "summary": "Üye Kaydı",
                "description": "Sisteme yeni kayıt açma",
                "operationId": "904da39f07955f15e8996ff12f00ce4d",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "mail"
                                    },
                                    "password": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/giris": {
            "post": {
                "tags": [
                    "Base API"
                ],
                "summary": "Üye girişi",
                "description": "Buradan üye girişi yapılır ve JWT token oluşturulur",
                "operationId": "7527b386226d0b8e3696c4ad37bc2b9b",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "mail"
                                    },
                                    "password": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Giriş başarılı",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/cikis": {
            "post": {
                "tags": [
                    "Base API"
                ],
                "summary": "Üye çıkışı",
                "description": "Buradan üye çıkışı yapılır ve JWT token silinir",
                "operationId": "a9cdbb79485b0f9715d558fa36ae7391",
                "responses": {
                    "200": {
                        "description": "Çıkış başarılı",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Token Bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profil": {
            "get": {
                "tags": [
                    "Base API"
                ],
                "summary": "Profil",
                "description": "Bu istek aktif olarak oturum açmış user'ın bilgilerini döndürür.",
                "operationId": "cbc4c1b018ce2cd64ef72ed5c434d3ef",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Token Bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sms": {
            "get": {
                "tags": [
                    "SMS API"
                ],
                "summary": "SMS Raporlarını Listeler",
                "operationId": "2f3d36d65ceaca2fbaefa884ffdb3fc7",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Tarih aralığının başlangıcı",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Tarih aralığının sonu",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail"
                    },
                    "404": {
                        "description": "fail"
                    },
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SMS API"
                ],
                "summary": "SMS kaydı",
                "description": "Buradan sms gönderimi ve kaydı gerçekleşir",
                "operationId": "d22807d27c35c5b93d241b98fe133338",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "number",
                                    "message"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "text"
                                    },
                                    "number": {
                                        "type": "tel"
                                    },
                                    "message": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sms/{id}": {
            "get": {
                "tags": [
                    "SMS API"
                ],
                "summary": "SMS rapor detayını getirir",
                "operationId": "9f0edba99dbbd7dfe8cd925a382879b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "SMS id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail"
                    },
                    "404": {
                        "description": "fail"
                    },
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}